Refactor namespaces and remove unused usings; add new project for extensions
