+++ src/Moneyman.Persistence/BankAccountRepository.cs
@@ -1,10 +1,7 @@
 
 using Moneyman.Interfaces;
-using System.Collections.Generic;
-using System.Linq;
 using Moneyman.Domain;
 using AutoMapper;
-using Microsoft.EntityFrameworkCore;
 
 namespace Moneyman.Persistence
 {
