+++ src/Moneyman.Persistence/PaydayRepository.cs
@@ -1,7 +1,5 @@
 
 using Moneyman.Interfaces;
-using System.Collections.Generic;
-using System.Linq;
 using Moneyman.Domain;
 using AutoMapper;
