+++ src/Moneyman.Services/Moneyman.Services.csproj
@@ -3,6 +3,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Moneyman.Domain\Moneyman.Domain.csproj" />
     <ProjectReference Include="..\Moneyman.Persistence\Moneyman.Persistence.csproj" />
+    <ProjectReference Include="..\Moneyman.Extentions\Moneyman.Extentions.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -12,7 +13,7 @@
     <PackageReference Include="Serilog.Sinks.Console" Version="4.1.0" />
     <PackageReference Include="Serilog.Sinks.Seq" Version="5.2.2" />
     <PackageReference Include="Serilog" Version="2.12.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="6.1.0" />    
+    <PackageReference Include="Serilog.AspNetCore" Version="6.1.0" />
   </ItemGroup>
 
   <PropertyGroup>
