+++ src/Moneyman.Tests/RepositoryTests/PaydayRepositoryTests.cs
@@ -16,7 +16,7 @@ using AutoMapper;
 using Moneyman.Domain.MapperProfiles;
 using Snapper;
 
-namespace Tests
+namespace Moneyman.Tests.RepositoryTests
 {
     [TestClass]
     public class PaydayRepositoryTests
@@ -26,30 +26,30 @@ namespace Tests
         private List<Payday> _paydays;
         private PaydayRepository NewPaydayRepository() =>
             new PaydayRepository(_contextMock.Object, _mapper);
-        
+
         [TestInitialize]
         public void SetUp()
         {
-            _contextMock = new  Mock<MoneymanContext>(); 
-            
+            _contextMock = new  Mock<MoneymanContext>();
+
             _paydays = new List<Payday>
             {
                 new Payday {Date = new DateTime(2022,1,1)},
                 new Payday {Date = new DateTime(2022,2,1)}
             };
-            
+
             var paydayMockDbSet = _paydays.AsQueryable().BuildMockDbSet();
 
-            _contextMock.Setup(x => x.Set<Payday>()).Returns(paydayMockDbSet.Object); 
+            _contextMock.Setup(x => x.Set<Payday>()).Returns(paydayMockDbSet.Object);
 
             var mappingConfig = new MapperConfiguration(mc =>
             {
                 mc.AddProfile(new TransactionDtoToTransactionProfile());
                 mc.AddProfile(new TransactionProfile());
             });
-            
+
             IMapper mapper = mappingConfig.CreateMapper();
-            _mapper = mapper;           
+            _mapper = mapper;
         }
 
         [TestMethod]
@@ -57,7 +57,7 @@ namespace Tests
         {
             _contextMock.Setup(x => x.Set<Payday>()).Returns(new List<Payday>{}.AsQueryable().BuildMockDbSet().Object);
             var repository = NewPaydayRepository();
-            var result = repository.GetAll();               
+            var result = repository.GetAll();
             result.Count().Should().Be(0);
         }
 
@@ -65,13 +65,13 @@ namespace Tests
         public void GetAll_WhenOneResult_ReturnsOneResult()
         {
             var repository = NewPaydayRepository();
-            var result = repository.GetAll();               
+            var result = repository.GetAll();
             result.Count().Should().Be(2);
         }
 
-        [TestMethod] 
+        [TestMethod]
         public async Task Add_WithMultiplePaydays_ReturnsSuccess()
-        {                
+        {
             List<Payday> updatedPaydays = null;
             using (var context = new MoneymanContext(BuildGenerateInMemoryOptions()))
             {
@@ -81,13 +81,13 @@ namespace Tests
                     paydayRepository.Add(payday);
                 }
                 await paydayRepository.Save();
-                
+
                 updatedPaydays = context.Paydays.ToList();
             }
 
             updatedPaydays.Should().NotBeNull();
             updatedPaydays.Count.Should().Be(_paydays.Count);
-            
+
             updatedPaydays.ShouldMatchSnapshot();
         }
