+++ src/Moneyman.Tests/ServiceTests/TransactionGeneratorTests/DtpMonthlyGenerationIntegrationTests.cs
@@ -16,7 +16,7 @@ using Microsoft.Extensions.Logging;
 using Moneyman.Services.Interfaces;
 using AutoMapper;
 
-namespace Moneyman.Tests
+namespace Moneyman.Tests.ServiceTests.TransactionGeneratorTests
 {
     [TestClass]
     public class DtpMonthlyGenerationIntegrationTests
