+++ src/Moneyman.Tests/ServiceTests/TransactionGeneratorTests/GenerateallTests.cs
@@ -13,7 +13,7 @@ using Microsoft.Extensions.Logging;
 using Moneyman.Services.Interfaces;
 using AutoMapper;
 
-namespace Moneyman.Tests
+namespace Moneyman.Tests.ServiceTests.TransactionGeneratorTests
 {
     [TestClass]
     public class GenerateAllTests
