+++ src/Moneyman.Tests/ServiceTests/TransactionServiceTests.cs
@@ -18,7 +18,7 @@ using Snapper;
 using Microsoft.Extensions.Logging;
 using AutoFixture;
 
-namespace Tests
+namespace Moneyman.Tests.ServiceTests
 {
     [TestClass]
     public class TransactionServiceTests
